Updated artikel, dienblad, kantine, kantinesimulatie_1, kassarij

+++ KantineSimulatieBasis/src/main/java/Artikel.java
@@ -20,7 +20,7 @@ public class Artikel {
         return naam;
     }
 
-    public int getPrijs(){
+    public double getPrijs(){
         return prijs;
     }
